Renamed shared node file
