+++ packages/plugins/intent/shared/src/nodes/IntentSearch.ts
